+++ Assets/Scripts/Gameplay/EndTurn.cs
@@ -7,12 +7,13 @@ public class EndTurn : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        
+        // pull turn counters for every unit
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+        // reduce every turn counter by X until one hits 0; use keyvaluepair for each unit to keep track? 
+        // assign unit turn
     }
 }
