+++ Assets/Scripts/Terrain/PathableTerrain.cs
@@ -4,6 +4,9 @@ using UnityEngine;
 
 public class PathableTerrain : MonoBehaviour
 {
+    // regular movement? difficult movement? add to turn counter? reduce move by X instead of block?
+    // LOS block bool?
+
     // Start is called before the first frame update
     void Start()
     {
