+++ Assets/Scripts/Terrain/UnpathableTerrain.cs
@@ -4,6 +4,9 @@ using UnityEngine;
 
 public class UnpathableTerrain : MonoBehaviour
 {
+    // pathable = false
+    // LOS block bool?
+
     // Start is called before the first frame update
     void Start()
     {
