+++ Assets/Scripts/Units/Move-MeleeUnit.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Move : MonoBehaviour
+{
+    // Turn counter per square moved or per move action (bool? + counter)
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        // Get the map size
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        // A* algorithm When movement is selected to get movement area. 
+        // Display movement area
+        // A* algorithm to show path when movement square is selected
+        // Accept? w/e
+    }
+}
